mcp_servers/vegalite_server/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
mcp-server==0.1.0
altair==4.2.2
vega==3.6.0
vega-datasets==0.9.0
pandas==1.5.3
pillow==9.5.0
python-dotenv==1.0.0
